Add Peregrine (#3)
* Add peregrine

* Use Peregrine to render the app

- Since store creation is moving to Peregrine, remove associated
code from this package and simplify what's left.
- Remove redux-actions dependency
- Change lint script to prettier

* Add Page component

Define a new component, Page, to be the dumb component rendered
by App, a smart component.

Also define a selector in the navigation reducer module that can
be used in `mapStateToProps` to provide initial state.

* Remove peregrine from webpack config whitelist

* Store paths as env vars

Use `dotenv` to load the contents of `.env` into `process.env`.
Paths such as the theme location and dev server public path may be
stored in `.env`.

* Update gitignore

* Switch peregrine to npm dependency