Add service worker (#13)
* Add and configure workbox

Use a new dependency, Workbox, to generate a service worker. Its
configuration is specified in `workbox-cli-config.js`.

Service workers impose new constraints on this app:

- Dev server must serve content over HTTPS
- Dev server can't serve the worker (cross-origin disallowed)

The output directory is now `web`, which M2 publishes. The worker
must be published this way. For now, clear the cache whenever you
update the worker.

* Add merge driver

* Add workbox plugin to webpack config

Also, consolidate calls to EnvironmentPlugin.

* Revert unintended change

* Fixup: prettier

* Fixup: lint

* Remove workbox-cli