(Breaking Change) Change Peregrine's bootstrapping API (#45)
<!-- (REQUIRED) What is the nature of this PR? -->
## This PR is a:
[ ] New feature
[ ] Enhancement/Optimization
[X] Refactor
[ ] Bugfix
[ ] Test for existing code
[ ] Documentation

<!-- (REQUIRED) What does this PR change? -->
## Summary

This PR changes the public API for bootstrapping a Peregrine application. Currently, `Peregrine` abstracts away the call to `ReactDOM.render`, and only accepts the DOM element that should host the application.

This is not ideal, because it makes it more difficult to wrap additional context providers in the root of the component tree.

This PR inverts things. The current API accepts all the data it needs, and then takes care of rendering on your behalf. The new API in this PR accepts all the data it needs, but returns to the consumer everything that would need to be used with `ReactDOM.render`. This is similar to switching from an HOC to a render prop - we're returning control back to the user of the component.

**Current Theme Code**
```js
const app = new Peregrine({
    apiBase: new URL('/graphql', location.origin).toString(),
    __tmp_webpack_public_path__: __webpack_public_path__
});

app.mount(document.getElementById('root'));
```

**New Theme Code**
```js
const { Provider } = bootstrap({
    apiBase: new URL('/graphql', location.origin).toString(),
    __tmp_webpack_public_path__: __webpack_public_path__
});

ReactDOM.render(<Provider />, document.getElementById('root'));
```

Note that the API where you call a function and it returns an object with a `Provider` is very similar to the `React.createContext` API in React 16. This is intentional.

Now, if a consumer wants to wrap additional providers at the root of the component tree, they can just do:

```js
ReactDOM.render(
   <SomeOtherProvider>
	 <PeregrineProvider />
   </SomeOtherProvider>,
   document.getElementById('root')
);
```

<!-- (OPTIONAL) What other information can you provide about this PR? -->
## Additional information

<!--
Thank you for your contribution!

Before submitting this pull request, please make sure you have read our Contribution Guidelines and your PR meets our contribution standards:
https://github.com/magento-research/peregrine/blob/develop/.github/CONTRIBUTION.md

Please fill out as much information as you can about your PR to help speed up the review process.
If your PR addresses an existing GitHub Issue, please refer to it in the title or Additional Information section to make the connection.

We may ask you for changes in your PR in order to meet the standards set in our Contribution Guidelines. PR's that do not comply with our guidelines may be closed at the maintainers' discretion.

Feel free to remove this section before creating this PR.
-->