Refactor the Page component and its children (#39)
* Refactor the Page component and its children

To accommodate drawers and other top-level elements, the page
components needs to be refactored. Also, the introduction of the
router broke the navigation, since it was being wired by the App
component, which is now defunct.

This commit uses a new global reducer, `app`, replacing the old
`navigation` one. It will manage drawer and overlay state.

The header and footer have been moved into the Main component.
This will preempt some stacking context issues.

A mask has been added to the app when a drawer is open. Further
masking may be added in the future for a modal dialog component.

* Update proptypes for triggers