WIP (DO NOT MERGE): New dev server concept
Checking in this WIP (which is not presently working) because it
contains concepts we will want to bring forward, but implemented in a
way that is probably unsustainable.

In order:
- Updated npm scripts so that common tasks were easier from package
 root
- Began effort to switch from Magento rendered app shell to
 middle-tier-rendered app shell.
- Began with HtmlWebpackPlugin and its extensions.
- Found incompatibilities with `webpack-dev-server` in desired
  approach.
- Began switch to `webpack-serve` over the unmaintained
  `webpack-dev-server`.
- Began switch to Webpack 4 to maintain compatibility with
`webpack-serve`. _(Maybe you see where this is going.)_
- Switched from deprecated CommonsChunkPlugin to `splitChunks`
configuration.
- Tweaked `splitChunks` configuration and HtmlWebpackPlugin
configuration to fix a topological sort issue caused by
MagentoRootComponentsPlugin.
- Refactored MagentoRootComponentsPlugin to be compatible with Webpack 4
compilation API. (Thanks to @adrien-louis-r whose commits I
cherry-picked, with attribution, to begin work on the plugin.)
- Discovered root components are fundamentally incompatible with Webpack4
with splitChunks in use.
- Determined that the RootComponents manifest strategy might not be
necessary at build time, if the middle tier specification can instead
select entry points based on preload.
- Decided to "park" this webpack-dev-mode upgrade effort for now.

Opened this as a PR so that it's easy to comment on and visible to
collaborators and community, but please do not merge this. It would be
very bad to merge this.