Move devdocs to package root, simplify scripts (#397)
* docs: add #393 content to devdocs and update

* chore: move devdocs to root to simplify scripts

Confirmed with @jcalcabemn that pwa-devdocs should not have its
dependencies versioned along with the rest of the Studio.

This helps us because it has external dependencies and defines a "build"
script, so every Lerna command had to specifically exclude it. Now our
Lerna scripts can be simpler.

Moved and audited package deps to compensate for change.