fix(ci): Fix CI deps at root. Fixes #883. (#884)
Description
===========

When migrating from npm to yarn, we tried to spread devDependencies
across the packages instead of centralizing them at root. Unfortunately,
we omitted a dependency that _needs_ to be at root so its CLI is
available in scripts,
[coveralls](https://github.com/nickmerwin/node-coveralls#readme).
Missing this dependency caused every build after the merge of #753
and all we need to do is add it back: `yarn add -W --dev coveralls`.

DangerCI has a transitive dependency on an outdated version of its
GitHub client library, but the semver dependency in DangerCI allows
updates, so we ran `yarn upgrade danger` and saw an Octokit update
appear in yarn.lock.