enhancement: unify fastly and local media URLs
Simplify the creation and consumption of image URLs, so they work the
same way in all environments.

This removees the extra 301 redirects to optimized images in production.

- Add a "preload" data to the HTML tag, which renders the Magento
backend URL as a data attribute if `USE_FASTLY` is true in the environment
- Modify `makeUrl()` to query that attribute (and therefore to find out at
runtime if Fastly is enabled), and use it as the base URL of an absolute
media URL when present
- Modify `makeUrl()` not to rewrite URLs for the needs of the onboard
image optimizer, but to normalize and append to them for the needs of
Fastly image optimization
- Modify the backend `addOptImgMiddleware()` feature instead, to be able
to read Fastly-style URLs.
- Remove image optimization logic from UPWARD entirely, since it is now
delegated to other layers.
- Update snapshots containing makeURL calls.