Fixed binary file reading
Fixed issue of 'binary' not being correctly parsed. Node.js parses 'binary' as 'latin1'. To achieve reading the file as a binary the file encoding needs to be set to null. See https://stackoverflow.com/a/46441727.

Removed buffers from needing to be turned into a string as it would corrupt the buffer.