Fixed binary file reading (#2050)
* Fixed binary file reading

Fixed issue of 'binary' not being correctly parsed. Node.js parses 'binary' as 'latin1'. To achieve reading the file as a binary the file encoding needs to be set to null. See https://stackoverflow.com/a/46441727.

Removed buffers from needing to be turned into a string as it would corrupt the buffer.

* Updated staticFromRoot file serving

Added encoding definition to be set to binary to properly serve favicon.ico.

Co-authored-by: AkivaTreestone <50888134+AkivaTreestone@users.noreply.github.com>
Co-authored-by: Revanth Kumar Annavarapu <35203638+revanth0212@users.noreply.github.com>