Update cms.js to use CSS Modules
Includes removing an import of the prop type `number`, which wasn't being used anywhere.
Elected to only accept `heading` as a property in the `classes` prop as that is the only class that is going to be used. Even though there is a cms.css file in the same directory, it wasn't being used previously, therefore I don't know what its purpose is, so I'm just playing it safe and leaving the code base as unchanged from my direct change.