Fix RootComponentsPlugin to be IE compatible (#2169)
* Fix RootComponentsPlugin to be IE compatible

As the final string is put together by calling `toString()` on variables that are functions, this means that the function itself is converted to a string, and then concatenated to the 1importerFactory` variable. The problem with this is that it's not transpiled, which means arrow functions and default parameters were making it to IE, causing the white screen of death.
This update doesn't fix all of the IE problems, i.e. there is still a white screen of deatch, but it does fix the errors caused by this file, and is intended as a "progressive bug fix"

* Fix code style with prettier

* Remove argument mutation and return `import` result

Co-authored-by: Devagouda <40405790+dpatil-magento@users.noreply.github.com>