feat: core team can scaffold new modules quick
As the BuildBus extension framework gains functionality, new features of Venia should be implemented as separate modules as much as possible.

This is the way of Magento, and it serves what partners desperately need:

 - A less monolithic, more maintainable Venia
 - Mix and match functionality
 - A Magento-like ability to enhance store functionality with a simple `npm install`
 - Examples of extensions, for extension developers to follow, so they start building/porting extensions and growing our ecosystem
 - Marketplace integration so extension developers can sell their stuff

For now, the script is intended for core contributor use only; use it if you want to add a package to the PWA Studio monorepo.

It creates a new module, adds it to the workspaces, and configures it for unit testing.

From the repo root directory:

```sh
node scripts/make-new-package.js
```

Answer the questionnaire and behold!