fix: correct regression in scaffolding from #2618
The Buildpack `createProject` utility was only linking locally to PWA
Studio packages that it specifically listed in its package.json. Other
packages that were transitive dependencies were not detected and linked.

That caused a
[regression](https://github.com/magento/pwa-studio/pull/2618#issuecomment-676521847) in #2618.

This PR alters the scaffolding code so it adds explicit dependencies for
every local dev-branch package. I also simplified and commented the
code as I was auditing it.

I also added some unit test coverage so that I could do more robust
testing on `createProject`.