Update size of Breadcrumbs container to allow focus ring to be visible
Using padding to "position" the breadcrumbs gives the focus ring space to actually render itself.
Adding extra padding to the bottom of the component allows for the bottom of the focus ring to be visible, and to compensate for that added height (and to ensure we're not pushing the rest of the UI down as a result), we negate that padding by using a negative margin.