Image not showing
Image tag from pagebuilder won't show in frontend. 
Image tag has "data-set" attribute when getting from pagebuilder, it is "src" attribute written the code file. 
We are getting wrong attribute, so we had to override this "configAggregator.js" to change ".getAttribute("src")" to ".getAttribute("data-src")" for image tag to show.